feat: allow guild-specific github issues
