+++ src/byte/lib/common/colors.py
@@ -5,8 +5,13 @@
 from typing import Final
 
 # --- Colors
-litestar_blue: Final = 0x7289DA
+litestar_blue: Final = 0x202235
+litestar_yellow: Final = 0xEDB641
+
 python_blue: Final = 0x4B8BBE
 python_yellow: Final = 0xFFD43B
+
 astral_yellow: Final = 0xD7FF64
 astral_purple: Final = 0x261230
+
+byte_teal: Final = 0x42B1A8
