+++ src/byte/views/__init__.py
@@ -1,7 +1,10 @@
 """Views for the bot."""
 
-from byte.views import forums
+from byte.views import abstract_views, astral, forums, python
 
 __all__ = [
+    "abstract_views",
+    "astral",
     "forums",
+    "python",
 ]
