+++ src/byte/views/abstract_views.py
@@ -1,20 +1,22 @@
 """Inheritable views that include extra functionality for base Views classes."""
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from copy import deepcopy
+from typing import TYPE_CHECKING, Any, Literal, TypedDict
 
-from discord import ButtonStyle, Embed, Interaction
+from discord import ButtonStyle, Colour, Embed, Interaction
 from discord.ui import Button, View, button
 
 if TYPE_CHECKING:
+    from datetime import datetime
     from typing import Self
 
     from discord.ext.commands import Bot
 
-__all__ = ("BaseEmbedView",)
+__all__ = ("ExtendedEmbed", "Field", "ButtonEmbedView")
 
 
-class BaseEmbedView(View):
+class ButtonEmbedView(View):
     """Base view including common buttons."""
 
     def __init__(self, author: int, bot: Bot, original_embed: Embed, minified_embed: Embed, *args, **kwargs) -> None:
@@ -86,3 +88,90 @@ class BaseEmbedView(View):
             _: Button object.
         """
         await self.learn_more_button_callback(interaction)
+
+
+class Field(TypedDict):
+    """Field type for ``ExtendedEmbed``.
+
+    .. note:: types are matching the ones in ``Embed.add_fields``.
+    """
+
+    name: Any
+    value: Any
+    inline: bool
+
+
+class ExtendedEmbed(Embed):
+    """Extended Embed class for discord.py."""
+
+    def add_field_dict(self, field: Field) -> Self:
+        """Add a field to the embed.
+
+        Args:
+            field (Field): The field to add to the embed.
+
+        Returns:
+            Self: The embed with the field added.
+        """
+        self.add_field(**field)
+        return self
+
+    def add_field_dicts(self, fields: list[Field]) -> Self:
+        """Add multiple fields to the embed.
+
+        Args:
+            fields (list[Field]): A list of fields to add to the embed.
+
+        Returns:
+            Self: The embed with the fields added.
+        """
+        for field in fields:
+            self.add_field_dict(field)
+        return self
+
+    @classmethod
+    def from_field_dicts(
+        cls,
+        colour: int | Colour | None = None,
+        color: int | Colour | None = None,
+        title: Any | None = None,
+        type: Literal["rich", "image", "video", "gifv", "article", "link"] = "rich",  # noqa: A002
+        url: Any | None = None,
+        description: Any | None = None,
+        timestamp: datetime | None = None,
+        fields: list[Field] | None = None,
+    ) -> Self:
+        """Create an embed from a list of fields.
+
+        Args:
+            colour (int | Colour | None): The colour of the embed.
+            color (int | Colour | None): The colour of the embed.
+            title (Any | None): The title of the embed.
+            type (Literal["rich", "image", "video", "gifv", "article", "link"]): The type of the embed.
+            url (Any | None): The URL of the embed.
+            description (Any | None): The description of the embed.
+            timestamp (datetime | None): The timestamp of the embed.
+            fields (list[Field] | None): A list of fields to add to the embed.
+
+        Returns:
+            Self: The embed with the fields added.
+        """
+        embed = cls(
+            colour=colour,
+            color=color,
+            title=title,
+            type=type,
+            url=url,
+            description=description,
+            timestamp=timestamp,
+        )
+        embed.add_field_dicts(fields or [])
+        return embed
+
+    def deepcopy(self) -> Self:
+        """Create a deep copy of the embed.
+
+        Returns:
+            Self: A deep copy of the embed.
+        """
+        return deepcopy(self)
