+++ src/byte/views/astral.py
@@ -2,12 +2,12 @@
 from __future__ import annotations
 
 from byte.lib.log import get_logger
-from byte.views.abstract_views import BaseEmbedView
+from byte.views.abstract_views import ButtonEmbedView
 
 __all__ = ("RuffView",)
 
 logger = get_logger()
 
 
-class RuffView(BaseEmbedView):
+class RuffView(ButtonEmbedView):
     """View for the Ruff embed."""
