+++ src/byte/views/forums.py
@@ -3,10 +3,9 @@
 from discord import ButtonStyle, Interaction
 from discord.ext.commands import Bot
 from discord.ui import Button, View, button
-from sqlalchemy import select
 
 from byte.lib.log import get_logger
-from server.domain.db.models import Guild
+from server.domain.guilds.dependencies import provides_guilds_service
 from server.lib.db import config
 
 __all__ = ("HelpThreadView",)
@@ -33,9 +32,8 @@ class HelpThreadView(View):
         # noinspection PyBroadException
         try:
             async with config.get_session() as session:
-                stmt = select(Guild).where(Guild.guild_id == self.guild_id)
-                result = await session.execute(stmt)
-                guild_settings = result.scalars().first()
+                guilds_service = await anext(provides_guilds_service(db_session=session))
+                guild_settings = await guilds_service.get(self.guild_id, id_attribute="guild_id")
 
             if guild_settings and guild_settings.github_config:
                 guild_repo = guild_settings.github_config.github_repository
@@ -44,7 +42,6 @@ class HelpThreadView(View):
                 )
             else:
                 logger.warning("no github configuration found for guild %s", self.guild_id)
-                await self.author.send("No GitHub configuration found for this guild. Please contact an admin.")
         except Exception:
             logger.exception("failed to setup view for guild %s", self.guild_id)
 
@@ -79,6 +76,7 @@ class HelpThreadView(View):
                 "invoking solve command for %s by %s on thread %s", ctx.channel, interaction.user, interaction.channel
             )
 
+            # noinspection PyBroadException
             try:
                 await solve_command.invoke(ctx)
                 await interaction.followup.send("Marked as solved and closed the help forum!", ephemeral=True)
