+++ src/byte/views/python.py
@@ -2,12 +2,12 @@
 from __future__ import annotations
 
 from byte.lib.log import get_logger
-from byte.views.abstract_views import BaseEmbedView
+from byte.views.abstract_views import ButtonEmbedView
 
 __all__ = ("PEPView",)
 
 logger = get_logger()
 
 
-class PEPView(BaseEmbedView):
+class PEPView(ButtonEmbedView):
     """View for the Python PEP embed."""
